Engine: Add DiagnosticText templates and BuildDiagnosticText() method
This adds a set of DiagnosticText templates and a BuildDiagnosticText()
method.

The BuildDiagnosticText() method returns a diagnostic text which will be
used to populate OperationOutcome's IssueComponent.Diagnostics.

The text varies based on the value of HttpStatusCode, currently we only
have templates for status codes 200 - OK and 201 - Created.
The text contains the full url to the resource. If the server supports
versioning the url will be 'resource/resource-id/_history/version-id',
otherwise 'resource/resource-id'.